# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Iraqi School Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface Admin {
  id: number;
  username: string;
  name: string;
  role: string;
}

export interface AuthResponse {
  token: string;
  admin: Admin;
}

export interface Student {
  id: number;
  fullName: string;
  studentCode: string;
  classId: number;
  className: string;
  gender: string;
  dateOfBirth: string;
  enrollmentDate: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  parentName?: string | null;
  /** @nullable */
  parentPhone?: string | null;
  /** @nullable */
  address?: string | null;
  status: string;
}

export interface StudentInput {
  fullName: string;
  classId: number;
  gender: string;
  dateOfBirth: string;
  phone?: string;
  parentName?: string;
  parentPhone?: string;
  address?: string;
  status?: string;
}

export interface StudentUpdate {
  fullName?: string;
  classId?: number;
  gender?: string;
  dateOfBirth?: string;
  phone?: string;
  parentName?: string;
  parentPhone?: string;
  address?: string;
  status?: string;
}

export interface Teacher {
  id: number;
  fullName: string;
  teacherCode: string;
  specialization: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  hireDate?: string;
  status: string;
}

export interface TeacherInput {
  fullName: string;
  specialization: string;
  phone: string;
  email?: string;
  address?: string;
  hireDate?: string;
  status?: string;
}

export interface TeacherUpdate {
  fullName?: string;
  specialization?: string;
  phone?: string;
  email?: string;
  address?: string;
  hireDate?: string;
  status?: string;
}

export interface Class {
  id: number;
  name: string;
  grade: string;
  teacherId: number;
  teacherName: string;
  capacity: number;
  studentCount: number;
  academicYear: string;
  /** @nullable */
  room?: string | null;
}

export interface ClassInput {
  name: string;
  grade: string;
  teacherId: number;
  capacity: number;
  academicYear: string;
  room?: string;
}

export interface ClassUpdate {
  name?: string;
  grade?: string;
  teacherId?: number;
  capacity?: number;
  academicYear?: string;
  room?: string;
}

export interface Subject {
  id: number;
  name: string;
  teacherId: number;
  teacherName: string;
  classId: number;
  className: string;
  weeklyHours: number;
  /** @nullable */
  description?: string | null;
}

export interface SubjectInput {
  name: string;
  teacherId: number;
  classId: number;
  weeklyHours: number;
  description?: string;
}

export interface SubjectUpdate {
  name?: string;
  teacherId?: number;
  classId?: number;
  weeklyHours?: number;
  description?: string;
}

export interface Attendance {
  id: number;
  studentId: number;
  studentName: string;
  classId: number;
  className: string;
  date: string;
  status: string;
  /** @nullable */
  notes?: string | null;
}

export interface AttendanceInput {
  studentId: number;
  classId: number;
  date: string;
  status: string;
  notes?: string;
}

export interface AttendanceUpdate {
  status?: string;
  notes?: string;
}

export interface Grade {
  id: number;
  studentId: number;
  studentName: string;
  subjectId: number;
  subjectName: string;
  classId: number;
  className: string;
  score: number;
  maxScore: number;
  examType: string;
  examDate: string;
  /** @nullable */
  notes?: string | null;
}

export interface GradeInput {
  studentId: number;
  subjectId: number;
  classId: number;
  score: number;
  maxScore: number;
  examType: string;
  examDate: string;
  notes?: string;
}

export interface GradeUpdate {
  score?: number;
  maxScore?: number;
  examType?: string;
  examDate?: string;
  notes?: string;
}

export interface DashboardStats {
  totalStudents: number;
  totalTeachers: number;
  totalClasses: number;
  totalSubjects: number;
  presentToday: number;
  absentToday: number;
  attendanceRate: number;
  averageGrade: number;
}

export interface AttendanceSummary {
  classId: number;
  className: string;
  presentCount: number;
  absentCount: number;
  lateCount: number;
  total: number;
}

export interface GradeDistribution {
  subjectName: string;
  averageScore: number;
  highestScore: number;
  lowestScore: number;
  passingCount: number;
  failingCount: number;
}

export interface Activity {
  id: number;
  type: string;
  description: string;
  timestamp: string;
  /** @nullable */
  entityName?: string | null;
}

export type ListStudentsParams = {
/**
 * @nullable
 */
classId?: number | null;
/**
 * @nullable
 */
search?: string | null;
};

export type ListTeachersParams = {
/**
 * @nullable
 */
search?: string | null;
};

export type ListAttendanceParams = {
/**
 * @nullable
 */
classId?: number | null;
/**
 * @nullable
 */
studentId?: number | null;
/**
 * @nullable
 */
date?: string | null;
};

export type ListGradesParams = {
/**
 * @nullable
 */
studentId?: number | null;
/**
 * @nullable
 */
classId?: number | null;
/**
 * @nullable
 */
subjectId?: number | null;
};