# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Iraqi School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        "401":
          description: Unauthorized
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List all classes
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /classes/{id}:
    get:
      operationId: getClass
      tags: [classes]
      summary: Get a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "404":
          description: Not found
    patch:
      operationId: updateClass
      tags: [classes]
      summary: Update a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: recordAttendance
      tags: [attendance]
      summary: Record attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendance/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /grades:
    get:
      operationId: listGrades
      tags: [grades]
      summary: List grades
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: subjectId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of grades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grade"
    post:
      operationId: createGrade
      tags: [grades]
      summary: Create a grade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
  /grades/{id}:
    patch:
      operationId: updateGrade
      tags: [grades]
      summary: Update a grade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
    delete:
      operationId: deleteGrade
      tags: [grades]
      summary: Delete a grade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [token, admin]
      properties:
        token: { type: string }
        admin:
          $ref: "#/components/schemas/Admin"
    Admin:
      type: object
      required: [id, username, name, role]
      properties:
        id: { type: integer }
        username: { type: string }
        name: { type: string }
        role: { type: string }
...
    Class:
      type: object
      required: [id, name, grade, teacherId, teacherName, capacity, studentCount, academicYear]
      properties:
        id: { type: integer }
        name: { type: string }
        grade: { type: string }
        teacherId: { type: integer }
        teacherName: { type: string }
        capacity: { type: integer }
        studentCount: { type: integer }
        academicYear: { type: string }
        room: { type: ["string", "null"] }
    ClassInput:
      type: object
      required: [name, grade, teacherId, capacity, academicYear]
      properties:
        name: { type: string }
        grade: { type: string }
        teacherId: { type: integer }
        capacity: { type: integer }
        academicYear: { type: string }
        room: { type: string }
    ClassUpdate:
      type: object
      properties:
        name: { type: string }
        grade: { type: string }
        teacherId: { type: integer }
        capacity: { type: integer }
        academicYear: { type: string }
        room: { type: string }
...
    Attendance:
      type: object
      required: [id, studentId, studentName, classId, className, date, status]
      properties:
        id: { type: integer }
        studentId: { type: integer }
        studentName: { type: string }
        classId: { type: integer }
        className: { type: string }
        date: { type: string }
        status: { type: string }
        notes: { type: ["string", "null"] }
    AttendanceInput:
      type: object
      required: [studentId, classId, date, status]
      properties:
        studentId: { type: integer }
        classId: { type: integer }
        date: { type: string }
        status: { type: string }
        notes: { type: string }
    AttendanceUpdate:
      type: object
      properties:
        status: { type: string }
        notes: { type: string }
    Grade:
      type: object
      required: [id, studentId, studentName, subjectId, subjectName, classId, className, score, maxScore, examType, examDate]
      properties:
        id: { type: integer }
        studentId: { type: integer }
        studentName: { type: string }
        subjectId: { type: integer }
        subjectName: { type: string }
        classId: { type: integer }
        className: { type: string }
        score: { type: number }
        maxScore: { type: number }
        examType: { type: string }
        examDate: { type: string }
        notes: { type: ["string", "null"] }
    GradeInput:
      type: object
      required: [studentId, subjectId, classId, score, maxScore, examType, examDate]
      properties:
        studentId: { type: integer }
        subjectId: { type: integer }
        classId: { type: integer }
        score: { type: number }
        maxScore: { type: number }
        examType: { type: string }
        examDate: { type: string }
        notes: { type: string }
    GradeUpdate:
      type: object
      properties:
        score: { type: number }
        maxScore: { type: number }
        examType: { type: string }
        examDate: { type: string }
        notes: { type: string }
...