# lib/api-zod/src/generated/types/attendance.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Iraqi School Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Attendance {
  id: number;
  studentId: number;
  studentName: string;
  classId: number;
  className: string;
  date: string;
  status: string;
  /** @nullable */
  notes?: string | null;
}