# lib/api-zod/src/generated/types/gradeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Iraqi School Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface GradeUpdate {
  score?: number;
  maxScore?: number;
  examType?: string;
  examDate?: string;
  notes?: string;
}